backend/requirements.txt
fastapi
uvicorn
pydantic
numpy
opencv-python
pillow
python-multipart
torch
torchvision
segment-anything
